refactor(api-server): move procedures and middlewares into specific files
