+++ packages/api-server/src/routers/auth/lud4.ts
@@ -1,4 +1,5 @@
-import { createTRPCRouter /* publicProcedure */ } from '../../trpc';
+import { createTRPCRouter } from '../../trpc';
+// import { publicProcedure } from '../../procedures';
 
 export const LUD4AuthRouter = createTRPCRouter({
   // TODO: Dolu?
