+++ packages/api-server/src/trpc.ts
@@ -1,19 +1,14 @@
-import { TRPCError, initTRPC } from '@trpc/server';
+import { initTRPC } from '@trpc/server';
 import type { inferAsyncReturnType } from '@trpc/server';
 import type { CreateExpressContextOptions } from '@trpc/server/adapters/express';
 import * as dotenv from 'dotenv';
-import { verify } from 'jsonwebtoken';
 import superjson from 'superjson';
 import { OpenApiMeta } from 'trpc-openapi';
 import { ZodError } from 'zod';
 
-import { JwtAuthTokenPayload } from '@sovereign-academy/types';
-
 import type { Dependencies } from './dependencies';
 
 dotenv.config();
-
-const AUTH_HEADER_KEY = 'authorization';
 
 /**
  * 1. CONTEXT
@@ -100,48 +95,5 @@ const t = initTRPC
 export const createTRPCRouter = t.router;
 export const mergeTRPCRouters = t.mergeRouters;
 
-/**
- * Public (unauthed) procedure
- *
- * This is the base piece you use to build new queries and mutations on your
- * tRPC API. It does not guarantee that a user querying is authorized, but you
- * can still access user session data if they are logged in
- */
-export const publicProcedure = t.procedure;
-
-/**
- * Reusable middleware that enforces users are logged in before running the
- * procedure
- */
-const enforceAuthenticatedUser = t.middleware(({ ctx, next }) => {
-  const authHeader = ctx.req.headers[AUTH_HEADER_KEY];
-  if (!authHeader || Array.isArray(authHeader))
-    throw new TRPCError({ code: 'UNAUTHORIZED' });
-
-  let payload: JwtAuthTokenPayload;
-  try {
-    payload = verify(
-      authHeader,
-      process.env['JWT_SECRET'] as string
-    ) as JwtAuthTokenPayload; // TODO validate payload with zod insteadƒ
-  } catch {
-    throw new TRPCError({ code: 'UNAUTHORIZED' });
-  }
-
-  return next({
-    ctx: {
-      user: payload,
-    },
-  });
-});
-
-/**
- * Protected (authed) procedure
- *
- * If you want a query or mutation to ONLY be accessible to logged in users, use
- * this. It verifies the session is valid and guarantees ctx.session.user is not
- * null
- *
- * @see https://trpc.io/docs/procedures
- */
-export const protectedProcedure = t.procedure.use(enforceAuthenticatedUser);
+export const createMiddleware = t.middleware;
+export const createProcedure = () => t.procedure;
